Update ob-core.el: Add safety check for goto-char
I've tried my own implementation of an `org-babel-execute:language` and got end error for the `goto-char` above. I'm not sure, what I would need to return from my own `org-babel-execute:language` in order to avoid that - the check above avoids that - and is hopefully a good enough fix to avoid such errors. If setting the position with `goto-char` is necessary, and `if` with an `else` branch would be a better fit - but I have no idea what to put into that branch.